\name{IsotopeR-package}
\alias{IsotopeR-package}
\docType{package}
\title{
IsotopeR package
}
\details{
\tabular{ll}{
Package: \tab IsotopeR\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2011-03-31\cr
License: GPL (>= 2)
LazyLoad: \tab yes\cr
}
}
\description{ The IsotopeR package is a mixing model for use in estimating diet contributions from food sources. The main functionality of the package is provided through the gui in the function IsotopeR. Other functions in the package are called by this interface, but could also be called by a user.
}
\author{
Jack Hopkins & Jake Ferguson

Maintainer: Jake Ferguson <troutinthemilk@gmail.com>
}
\references{
Hopkins, J. B., & Ferguson, J. M. (2012). Estimating the Diets of Animals Using Stable Isotopes and a Comprehensive Bayesian Mixing Model. PLoS ONE, 7(1), 1-13. doi:10.1371/journal.pone.0028478

}

