\name{IsotopeR}
\alias{IsotopeR}
\title{
IsotopeR
}
\description{
This runs the gui for running an IsotopeR model using JAGS.
}
\usage{
IsotopeR()
}
\references{
Hopkins JB III, Ferguson JM (2012) Estimating the Diets of Animals Using Stable Isotopes and a Comprehensive Bayesian Mixing Model. PLoS ONE 7(1)
}
\author{
Jake Hopkins, Jake Ferguson
}

\seealso{
\code{\link{IsoWrapper}}
}
\examples{
#to run IsotopeR call
\dontrun{IsotopeR()}
#then follow the gui prompts.

#type 'vignette("IsotopeR") for a full tutorial.
}

