\name{IsotopeRnoconcnodiscrimgroup}
\alias{IsotopeRnoconcnodiscrimgroup}
\docType{data}
\title{
No Concentration Dependence or Discrimination Error Model
}
\description{
The JAGS IsotopeR model  with group structure and with no concentration dependence or discrimination error. Called by IsotopeR().
}


