\name{IsotopeRnoconcnomenodiscrimgroup}
\alias{IsotopeRnoconcnomenodiscrimgroup}
\docType{data}
\title{
No Concentration Dependence, Measurement Error or Discrimination Error Model
}
\description{
The JAGS IsotopeR model  with group structure and with no concentration dependence, measurement error or discriminination error. Called by IsotopeR().
}

