\name{IsotopeRnoconcnomenoind}
\alias{IsotopeRnoconcnomenoind}
\docType{data}
\title{
No Concentration Dependence or Measurement Error Model
}
\description{
The JAGS IsotopeR model with no individual or group structure and with no concentration dependence or measurement error. Called by IsotopeR().
}


