\name{IsotopeR-package}
\alias{IsotopeR-package}
\docType{package}
\title{
IsotopeR package
}
\details{
\tabular{ll}{
Package: \tab IsotopeR\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2011-03-31\cr
License: GPL (>= 2)
LazyLoad: \tab yes\cr
}
}
\description{ The IsotopeR package is a mixing model for use in estimating diet contributions from food sources. The main functionality of the package is provided through the gui in the function IsotopeR. Other functions in the package are called by this interface, but could also be called by a user.
}
\author{
Jack Hopkins & Jake Ferguson

Maintainer: Jake Ferguson <troutinthemilk@gmail.com>
}
\references{
Hopkins and Ferguson, submitted

Semmens, B.X., Ward, E.J., Moore, J.W., and Darimont, C.T.: Quantifying Inter-and Intra-Population Niche Variability Using Hierarchical Bayesian Stable Isotope Mixing Models, Public Library of Science, 2009

}
\seealso{
\code{\link[siar]{siarmenu}}
}
