\name{IsotopeR}
\alias{IsotopeR}
\title{
IsotopeR
}
\description{
This runs the gui for running an IsotopeR model using JAGS.
}
\usage{
IsotopeR()
}
\references{
Hopkins and Ferguson, submitted
Semmens, B.X., Ward, E.J., Moore, J.W., and Darimont, C.T.: Quantifying Inter-and Intra-Population Niche Variability Using Hierarchical Bayesian Stable Isotope Mixing Models, Public Library of Science, 2009
}
\author{
Jake Hopkins, Jake Ferguson
}

\seealso{
\code{\link[siar]{siarmenu}}
\code{\link{IsoWrapper}}
}
\examples{
#to run IsotopeR call:
#IsotopeR()
#then follow the gui prompts.
}

